requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# GitHub App authentication
PyJWT==2.8.0
cryptography==42.0.2

# HTTP client
httpx==0.26.0
aiohttp==3.9.3

# OpenAI API
openai==1.12.0

# Configuration and environment
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Rate limiting
aiolimiter==1.1.0

# Retry logic
tenacity==8.2.3

# Logging
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
